Refactor async storage utility imports and add new utility functions for JSON extraction, date formatting, and error logging
