+++ src/screens/SkillsDashboardScreen.tsx
@@ -13,6 +13,7 @@ import {
   IdentifiedSkill,
 } from "../services/userSkillsService";
 import { SKILLS_TAXONOMY, mapSkillToTaxonomy } from "../services/skillTaxonomyService";
+import { formatDate } from "../utils/dateUtils";
 
 type SkillsDashboardScreenNavigationProp = StackNavigationProp<RootStackParamList>;
 type SkillsDashboardScreenRouteProp =
@@ -147,19 +148,6 @@ export default function SkillsDashboardScreen({ navigation, route: _route }: Rea
     }).length;
     const percentage = total > 0 ? Math.round((identified / total) * 100) : 0;
     return { identified, total, percentage };
-  };
-
-  const formatDate = (dateString: string): string => {
-    const date = new Date(dateString);
-    const now = new Date();
-    const diffMs = now.getTime() - date.getTime();
-    const diffDays = Math.floor(diffMs / (1000 * 60 * 60 * 24));
-
-    if (diffDays === 0) return "Today";
-    if (diffDays === 1) return "Yesterday";
-    if (diffDays < 7) return `${diffDays} days ago`;
-    if (diffDays < 30) return `${Math.floor(diffDays / 7)} weeks ago`;
-    return date.toLocaleDateString();
   };
 
   const categories = ["All", ...Object.keys(SKILLS_TAXONOMY)];
